CHECKPOINT: fix(logger): update stage test to handle zap's int64 type
The TestWithStage function in the logger package was failing due to a type
mismatch between the expected int type and zap's internal int64 representation
of numeric fields. This change updates the test to expect int64 values,
aligning with zap's internal type system.

Rather than fighting against zap's type conversion behavior, we adapt our
test to match the established pattern. This maintains compatibility while
keeping the stage.go implementation clean and idiomatic.

- Changed wantStage type in TestWithStage from int to int64
- Updated comments to explain the type choice
- Maintains existing behavior of the staging system
- No changes to production code required

+++ cmd/wfdevice/logger/logger_test.go
@@ -182,7 +182,7 @@ func TestWithStage(t *testing.T) {
 	tests := []struct {
 		name       string
 		inputStage int
-		wantStage  int
+		wantStage  int64 // Changed to int64 to match zap's internal representation
 	}{
 		{
 			name:       "normal stage",
